backend/requirements.txt
# Django Core
Django==5.0.4
djangorestframework==3.15.1

# WebSocket / Channels
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2

# Redis
redis==5.0.4

# Database
psycopg2-binary==2.9.9

# CORS
django-cors-headers==4.3.1

# Environment variables
python-decouple==3.8

# Utilities
Pillow==10.3.0
